requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
acme==2.7.1
    # via
    #   -r requirements.in
    #   certbot
alembic==1.12.0
    # via flask-migrate
alembic-autogenerate-enums==0.1.2
    # via -r requirements.in
amqp==5.1.1
    # via kombu
aniso8601==9.0.1
    # via flask-restful
arrow==1.3.0
    # via -r requirements.in
async-timeout==4.0.3
    # via redis
asyncpool==1.0
    # via -r requirements.in
attrs==23.1.0
    # via jsonlines
bcrypt==4.0.1
    # via
    #   flask-bcrypt
    #   paramiko
beautifulsoup4==4.12.2
    # via cloudflare
billiard==4.1.0
    # via celery
blinker==1.6.3
    # via
    #   flask
    #   flask-mail
    #   flask-principal
boto3==1.28.66
    # via -r requirements.in
botocore==1.31.66
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
celery[redis]==5.3.4
    # via -r requirements.in
certbot==2.7.1
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   requests
    #   sentry-sdk
certsrv[ntlm]==2.1.1
    # via -r requirements.in
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cloudflare==2.12.4
    # via -r requirements.in
configargparse==1.7
    # via certbot
configobj==5.0.8
    # via certbot
cryptography==41.0.4
    # via
    #   -r requirements.in
    #   acme
    #   certbot
    #   josepy
    #   paramiko
    #   pyopenssl
    #   pyspnego
    #   requests-ntlm
deprecated==1.2.14
    # via limits
distro==1.8.0
    # via certbot
dnspython==1.15.0
    # via dnspython3
dnspython3==1.15.0
    # via -r requirements.in
dyn==1.8.6
    # via -r requirements.in
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-cors
    #   flask-limiter
    #   flask-mail
    #   flask-migrate
    #   flask-principal
    #   flask-restful
    #   flask-sqlalchemy
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-cors==4.0.0
    # via -r requirements.in
flask-limiter==3.5.0
    # via -r requirements.in
flask-mail==0.9.1
    # via -r requirements.in
flask-migrate==4.0.5
    # via -r requirements.in
flask-principal==0.4.0
    # via -r requirements.in
flask-replicated==2.1
    # via -r requirements.in
flask-restful==0.3.10
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
future==0.18.3
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
hvac==1.2.1
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via
    #   certbot
    #   flask
importlib-resources==6.1.0
    # via limits
inflection==0.5.1
    # via -r requirements.in
itsdangerous==2.1.2
    # via
    #   -r requirements.in
    #   flask
javaobj-py3==0.4.3
    # via pyjks
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
josepy==1.13.0
    # via
    #   acme
    #   certbot
jsonlines==4.0.0
    # via cloudflare
kombu==5.3.2
    # via celery
limits==3.6.0
    # via flask-limiter
lockfile==0.12.2
    # via -r requirements.in
logmatic-python==0.1.7
    # via -r requirements.in
mako==1.2.4
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==2.21.0
    # via
    #   -r requirements.in
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.23.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
ndg-httpsclient==0.5.1
    # via -r requirements.in
ordered-set==4.1.0
    # via flask-limiter
packaging==23.2
    # via
    #   gunicorn
    #   limits
paramiko==3.3.1
    # via -r requirements.in
parsedatetime==2.6
    # via certbot
pem==23.1.0
    # via -r requirements.in
prompt-toolkit==3.0.39
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   ndg-httpsclient
    #   pyasn1-modules
    #   pyjks
    #   python-ldap
pyasn1-modules==0.3.0
    # via
    #   pyjks
    #   python-ldap
pycparser==2.21
    # via cffi
pycryptodomex==3.19.0
    # via pyjks
pygments==2.16.1
    # via rich
pyhcl==0.4.5
    # via hvac
pyjks==20.0.0
    # via -r requirements.in
pyjwt==2.8.0
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyopenssl==23.2.0
    # via
    #   -r requirements.in
    #   acme
    #   josepy
    #   ndg-httpsclient
pyrfc3339==1.1
    # via
    #   acme
    #   certbot
pyspnego==0.10.2
    # via requests-ntlm
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   celery
python-json-logger==2.0.7
    # via logmatic-python
python-ldap==3.4.3
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   acme
    #   certbot
    #   flask-restful
    #   pyrfc3339
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   cloudflare
redis==4.6.0
    # via
    #   -r requirements.in
    #   celery
requests==2.31.0
    # via
    #   -r requirements.in
    #   acme
    #   certsrv
    #   cloudflare
    #   hvac
    #   requests-ntlm
requests-ntlm==1.2.0
    # via certsrv
retrying==1.3.4
    # via -r requirements.in
rich==13.6.0
    # via flask-limiter
s3transfer==0.7.0
    # via boto3
sentry-sdk==1.32.0
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   configobj
    #   flask-restful
    #   python-dateutil
    #   retrying
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==1.3.24
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via -r requirements.in
tabulate==0.9.0
    # via -r requirements.in
twofish==0.3.0
    # via pyjks
types-python-dateutil==2.8.19.14
    # via arrow
typing-extensions==4.8.0
    # via
    #   alembic
    #   flask-limiter
    #   kombu
    #   limits
tzdata==2023.3
    # via celery
urllib3==1.26.18
    # via
    #   botocore
    #   requests
    #   sentry-sdk
validators==0.22.0
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.8
    # via prompt-toolkit
werkzeug==3.0.0
    # via
    #   -r requirements.in
    #   flask
wrapt==1.15.0
    # via deprecated
xmltodict==0.13.0
    # via -r requirements.in
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
